Added card template. Added search. Images.
Added card template for the initial recipe images/cards. Clicking a card
will result in a second window popping up that has the full sized image
as well as the actual recipe in it.

Search by title works nicely. Fades recipes in and out if they match the
criteria or not. A full text search on the recipe is possible, but lets
not do that yet. Too much would pop up for 'bak' I bet. Or 'heat' and
other general cooking terms. So title search will do for now.

Added some styling to force img's in cards to be a specific size. I
wonder if there's a way to fix it without using a background image?